Fuse fmpz_mpoly/get*
